backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.5
Django==5.0.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-ninja==1.1.0
django-ninja-extra==0.20.4
django-ninja-jwt==5.3.0
injector==0.21.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pygraphviz==1.12
PyJWT==2.8.0
pyparsing==3.1.1
sqlparse==0.4.4
typing_extensions==4.10.0
